scrapers/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
lxml>=4.9.0
psycopg2-binary>=2.9.0
pika==1.3.2
python-dotenv==1.0.0
schedule==1.2.0